Добавлено разрешение на использование местоположения. Добавлены функции для запроса погоды по координатам

+++ app/src/main/java/com/example/weatherApp/RestServices/DataParseClassForCurrentWeather.kt
@@ -41,7 +41,7 @@ data class DataParseClassForCurrentWeather(
 
     data class Main(
         @SerializedName("temp") val temp: Double,
-        @SerializedName("pressure") val pressure: Int,
+        @SerializedName("pressure") val pressure: Double,
         @SerializedName("humidity") val humidity: Int,
         @SerializedName("temp_min") val temp_min: Double,
         @SerializedName("temp_max") val temp_max: Double
